Some refactor Create readme about endpoints Add postman collection

+++ src/main/java/com/codecool/companymanager/service/DepartmentService.java
@@ -42,6 +42,10 @@ public class DepartmentService {
 
     @Transactional
     public void deleteById(long id) {
-        departmentRepository.deleteById(id);
+        if (departmentRepository.existsById(id)) {
+            departmentRepository.deleteById(id);
+        } else {
+            throw new NoSuchElementException();
+        }
     }
 }
